feature/route-refactor: Added routing on load
