+++ web/html/login.html
@@ -2,6 +2,7 @@
   <link rel="stylesheet" href="../css/login.css" />
   <script src="https://accounts.google.com/gsi/client" defer async></script>
   <script type="module" src="../scripts/login.js" defer></script>
+  <title>Spyder Login</title>
 </head>
 <body>
   <div class="container">
