+++ web/index.html
@@ -5,7 +5,6 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" href="./css/styles.css" />
 
-    <script type="module" src="./scripts/router.js" defer></script>
     <script type="module" src="./scripts/index.js" defer></script>
 
     <link rel="icon" type="image/x-icon" href="resources/SpyderwareLogo.png" />
