+++ web/scripts/index.js
@@ -1,11 +1,23 @@
-import { Routes } from "./utils.js";
+import { Routes, PATH_CHANGE_EVENT_NAME } from "./utils.js";
 import { changeRoute } from "./router.js";
+import { initAuth, isLoggedIn } from "./auth.js";
 
-window.addEventListener('DOMContentLoaded', loadLogin);
+window.addEventListener('DOMContentLoaded', init);
 
-/**
- * Load the login page when the user first accesses the site.
- */
-function loadLogin() {
-    changeRoute(Routes.Login, false);
+function init() {
+    if (!isLoggedIn()) {
+        initAuth();
+    } else if (window.location.pathname === "/") {
+        initRootPathPage();
+    } else {
+        routeToPage();
+    }
+}
+
+function initRootPathPage() {
+    changeRoute(Routes.Homepage, true);
+}
+
+function routeToPage() {
+    window.dispatchEvent(new Event(PATH_CHANGE_EVENT_NAME));
 }
